Updated loading State
